refactor: Account Module
