+++ app/build.gradle.kts
@@ -84,6 +84,7 @@ dependencies {
     implementation(projects.feature.savings)
     implementation(projects.feature.qr)
     implementation(projects.feature.transferProcess)
+    implementation(projects.feature.account)
