+++ feature/account/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    alias(libs.plugins.mifos.android.feature)
+    alias(libs.plugins.mifos.android.library.compose)
+}
+
+android {
+    namespace = "org.mifos.mobile.feature.account"
+}
+
+dependencies {
+    implementation(projects.ui)
+    implementation(projects.core.common)
+    implementation(projects.core.model)
+    implementation(projects.core.data)
+    implementation(libs.reactivex.rxjava2.android)
+    implementation(libs.reactivex.rxjava2)
+    api(libs.androidx.compose.material)
+    implementation("com.github.rahul-gill.mifos-ui-library:uihouse:alpha-2.1")
+    testImplementation(libs.junit)
+    androidTestImplementation(libs.androidx.test.ext.junit)
+    androidTestImplementation(libs.espresso.core)
+}
