+++ feature/account/src/main/res/values/colors.xml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="white">#ffffffff</color>
+    <color name="black">#000000</color>
+    <color name="background">#eaeaea</color>
+    <color name="deposit_green">#ff14c416</color>
+    <color name="light_green">#ff8bf98a</color>
+    <color name="light_yellow">#fff9ac06</color>
+    <color name="light_blue">#FF87DBF9</color>
+    <color name="light_red">#fff9393c</color>
+    <color name="light_grey">#ffd1d1d1</color>
+    <color name="transaction_detail_bg">#8ad3da44</color>
+    <color name="transaction_detail_table_header_bg">#8ada6134</color>
+    <color name="blue">#ff003fff</color>
+    <color name="red">#FF0000</color>
+    <color name="light_black">#1C1C1C</color>
+
+    <color name="dashboard_gray_background">#B2C1C8</color>
+
+    <!-- A light Holo shade of blue -->
+    <color name="blue_light">#ff33b5e5</color>
+    <!-- A light Holo shade of gray -->
+    <color name="gray_light">#33999999</color>
+    <!-- A dark Holo shade of gray -->
+    <color name="gray_dark">#BB666666</color>
+    <!-- A light Holo shade of green -->
+    <color name="green_light">#ff99cc00</color>
+    <!-- A light Holo shade of red -->
+    <color name="red_light">#ffff4444</color>
+    <!-- A dark Holo shade of blue -->
+    <color name="blue_dark">#ff0099cc</color>
+    <!-- A dark Holo shade of green -->
+    <color name="green_dark">#ff669900</color>
+    <!-- A dark Holo shade of red -->
+    <color name="red_dark">#ffcc0000</color>
+    <!-- A Holo shade of purple -->
+    <color name="purple">#ffaa66cc</color>
+    <!-- A light Holo shade of orange -->
+    <color name="orange_light">#ffffbb33</color>
+    <!-- A dark Holo shade of orange -->
+    <color name="orange_dark">#ffff8800</color>
+    <!-- A really bright Holo shade of blue -->
+    <color name="blue_bright">#ff00ddff</color>
+    <!-- A really bright Holo shade of gray -->
+    <color name="gray_bright">#33CCCCCC</color>
+
+    <!-- Material design colors -->
+
+    <!-- New UI color palette -->
+    <!--<color name="primary">#068799</color>-->
+    <!--<color name="primary_dark">#006978</color>-->
+    <!--<color name="primary_light">#56c8d8</color>-->
+    <color name="primary">#03A9F4</color>
+    <color name="primary_dark">#0288D1</color>
+    <color name="primary_light">#B3E5FC</color>
+    <color name="accent">#FF4081</color>
+    <color name="primary_text">#212121</color>
+    <color name="secondary_text">#757575</color>
+    <color name="icons">#FFFFFF</color>
+    <color name="divider">#BDBDBD</color>
+    <color name="cell_background_color">#ffffff</color>
+    <color name="header_line_color">@color/blue_light</color>
+    <color name="background_default_darker">#EEEEEE</color>
+    <color name="transparent">#00000000</color>
+
+    <!--Swipe Refresh Colors-->
+    <array name="swipeRefreshColors">
+        <item>@color/blue_light</item>
+        <item>@color/green_light</item>
+        <item>@color/red_light</item>
+        <item>@color/orange_light</item>
+    </array>
+
+</resources>
