+++ settings.gradle.kts
@@ -35,3 +35,4 @@ include(":feature:registration")
 include(":feature:savings")
 include(":feature:qr")
 include(":feature:transfer-process")
+include(":feature:account")
